// src/app/interfaces/productionrecords.ts
export interface IStandardinspect {
    id: number;
    productionRecordId: number;
    workerId: number;
    frontLeftWeight: number;
    frontCenterWeight: number;
    frontRightWeight: number;
    frontLeftThickness: number;
    frontCenterThickness: number;
    frontRightThickness: number;
    frontFabricWidth: number;
    frontFabricLength: number;
    frontFabricSet: number;
    frontIsLeftCrease: boolean;
    frontIsSmoothRoll: boolean;
    rearLeftWeight: number;
    rearCenterWeight: number;
    rearRightWeight: number;
    rearLeftThickness: number;
    rearCenterThickness: number;
    rearRightThickness: number;
    rearFabricWidth: number;
    rearFabricLength: number;
    rearFabricSet: number;
    rearIsLeftCrease: boolean;
    rearIsSmoothRoll: boolean;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
}